Add files via upload
In previous years, Landsat Collection 1 was used to make the SCT inputs. Now only Landsat Collection 2 is being updated. This code provides the method for producing SCT inputs using Collection 2. It does not loop through the years as the previous code did, because it is only intended to be used once each year.